backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.7.0
opencv-python>=4.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
google-genai>=0.1.0
elevenlabs>=0.2.0
requests>=2.25.0
Pillow>=8.0.0
pymongo>=4.0.0